requirements.txt
bs4
pychalk
html5lib
aiohttp